+++ App.vue
@@ -7,7 +7,7 @@
 	import {
 		version
 	} from './package.json'
-	import { UNI_ADMIN_THEME } from './store/constants.js'
+	import { UNI_ADMIN_THEME, UNI_ID_TOKEN_EXPIRED } from './store/constants.js'
 	export default {
 		created() {
 			this.clear = undefined
@@ -35,6 +35,15 @@
 				'background:#007aff ;padding: 1px; border-radius: 0 3px 3px 0;  color: #fff; font-weight: bold;'
 			)
 			// #endif
+			// #ifdef H5
+			// 此处为强制用户必须登录才能访问其他页面，如果需要部分页面不登录也能访问，需要在此过滤这些页面不执行下面的代码或直接注释掉此代码
+			// let uni_id_token_expired = uni.getStorageSync(UNI_ID_TOKEN_EXPIRED);
+			// if (!uni_id_token_expired || uni_id_token_expired < Date.now()) {
+			// 	uni.reLaunch({
+			// 		url: config.login.url
+			// 	})
+			// }
+			// #endif
 			// 线上示例使用
 			// console.log('%c uni-app官方团队诚邀优秀前端工程师加盟，一起打造更卓越的uni-app & uniCloud，欢迎投递简历到 hr2013@dcloud.io', 'color: red');
 			console.log('App Launch')
@@ -46,7 +55,7 @@
 			uni.$on('uni-id-pages-login-success', () => {
 				this.init()
 			})
-			
+
 			// theme
 			this.SET_THEME(uni.getStorageSync(UNI_ADMIN_THEME) || 'default')
 		},