+++ uni_modules/uni-id-pages/common/login-page.mixin.js
@@ -36,7 +36,6 @@ let mixin = {
 				})
 			}
 			this.$nextTick(n => {
-				console.log(this.$refs.uniFabLogin);
 				this.$refs.uniFabLogin.login({
 					code:e.code
 				}, 'weixin')