+++ uni_modules/uni-id-pages/components/cloud-image/cloud-image.vue
@@ -46,13 +46,10 @@
 		watch: {
 			src:{
 				handler(src) {
-					// console.log(src);
-					// console.log(src.substring(0, 8));
 					if (src&&src.substring(0, 8) == "cloud://") {
 						uniCloud.getTempFileURL({
 							fileList: [src]
 						}).then(res=>{
-							// console.log(res);
 							this.cSrc = res.fileList[0].tempFileURL
 						})
 					}else{