+++ uni_modules/uni-id-pages/pages/common/webview/webview.vue
@@ -20,7 +20,6 @@
 				});
 				title = "页面路径错误"
 			}else{
-				console.log(url,title);
 				this.url = url;
 			}
 			if(title){