+++ uni_modules/uni-id-pages/pages/login/login-smscode.vue
@@ -38,7 +38,6 @@
 			this.phone = phoneNumber;
 		},
 		onShow() {
-			console.log('onShow');
 			// #ifdef H5
 			document.onkeydown = event => {
 				var e = event || window.event;
@@ -59,7 +58,8 @@
 					this.$refs.smsCode.focusSmsCodeInput = true
 					return uni.showToast({
 						title: '验证码不能为空',
-						icon: 'none'
+						icon: 'none',
+						duration: 3000
 					});
 				}
 				uniIdCo.loginBySms({
@@ -67,7 +67,6 @@
 					"code": this.code,
 					"captcha": this.captcha
 				}).then(e => {
-					console.log(e);
 					this.loginSuccess(e)
 				}).catch(e => {
 					if (e.errCode == 'uni-id-captcha-required') {