+++ uni_modules/uni-id-pages/pages/userinfo/change_pwd/change_pwd.vue
@@ -82,8 +82,6 @@
 			 * 完成并提交
 			 */
 			submit() {
-				console.log("formData", this.formData);
-				console.log('rules', this.rules);
 				this.$refs.form.validate()
 					.then(res => {
 						let {
@@ -94,7 +92,6 @@
 								oldPassword,
 								newPassword
 							}).then(e => {
-								console.log(e);
 								uni.removeStorageSync('uni_id_token');
 								uni.setStorageSync('uni_id_token_expired', 0)
 								uni.redirectTo({
@@ -109,7 +106,6 @@
 					}).catch(errors => {
 						let key = errors[0].key
 						key = key.replace(key[0], key[0].toUpperCase())
-						console.log(key, 'focus' + key);
 						this['focus' + key] = true
 					})
 			}