+++ uni_modules/uni-id-pages/uniCloud/cloudfunctions/uni-id-co/config/permission.js
@@ -74,5 +74,8 @@ module.exports = {
   },
   unbindApple: {
     auth: true
+  },
+  setPwd: {
+    auth: true
   }
 }