+++ uni_modules/uni-id-pages/uniCloud/cloudfunctions/uni-id-co/middleware/index.js
@@ -3,5 +3,6 @@ module.exports = {
   uniIdLog: require('./uni-id-log'),
   validate: require('./validate'),
   accessControl: require('./access-control'),
+  verifyRequestSign: require('./verify-request-sign'),
   ...require('./rbac')
 }