+++ uni_modules/uni-id-pages/uniCloud/cloudfunctions/uni-id-co/module/multi-end/utils.js
@@ -19,7 +19,9 @@ async function isAuthorizeApproved ({
       errCode: ERROR.ACCOUNT_NOT_EXISTS
     }
   }
-  const userMatched = await findUser({
+  const {
+    userMatched
+  } = await findUser({
     userQuery: userRecord,
     authorizedApp: appIdList
   })